Move the Ingest/IngestUpdate logic out of the IngestTracker
